allow derived children of BarController to set label position as a width fraction, not an absolute position
git-svn-id: svn://localhost/ardour2/branches/3.0@8597 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/automation_controller.h
@@ -55,7 +55,7 @@ public:
 
 private:
 	AutomationController (boost::shared_ptr<ARDOUR::Automatable> parent, boost::shared_ptr<ARDOUR::AutomationControl> ac, Gtk::Adjustment* adj);
-	std::string get_label (int&);
+	std::string get_label (double&);
 
 	void start_touch();
 	void end_touch();
